+++ src/multi-tenancy/contracts/future-multi-tenancy-evolution-profile.js
@@ -1,0 +1,53 @@
+export class FutureMultiTenancyEvolutionProfile {
+  constructor({
+    profileName,
+    evolutionPrinciples = [], stableArchitecturalInvariants = [], evolutionDrivers = [], evolutionHorizons = [],
+    cellModelRequirements = [], cellDesignOptions = [], placementEvaluationFactors = [], placementRebalancingSteps = [],
+    progressiveIsolationRequirements = [], sovereignDeploymentElements = [], providerPortabilityImprovements = [],
+    futurePortabilityPackageContents = [], continuousVerificationCapabilities = [], digitalTwinFields = [],
+    policyAsCodeMaturityItems = [], adaptiveGovernanceTargets = [], predictiveCapacityInputs = [], privacyPreservingControls = [],
+    federatedLearningGovernanceAreas = [], aiAssistedOperationsAreas = [], remediationDefinitionFields = [],
+    selfServiceEvolutionItems = [], ecosystemGrowthAreas = [], collaborationEvolutionFields = [],
+    identityEvolutionPreservedElements = [], migrationFactoryMovementTypes = [], deletionVerificationDistinctions = [],
+    sustainabilityConsiderations = [], evolutionGovernanceSteps = [], evolutionMetrics = [], evolutionRisks = [],
+    architecturalRules = [],
+    tenantSemanticsStableAcrossEvolution = true, evolutionIncrementalReversibleEvidenceDriven = true,
+    automationNarrowsRiskNotHidesDecisions = true, isolationImprovementsPortable = true,
+    dedicatedSovereignReuseCommonContracts = true, migrationIsProductCapability = true, portabilityIncludesEvidence = true,
+    aiOperationsBoundedByAccountability = true, privacyInsightsNoImplicitRights = true, optimizationCannotWeakenCommitments = true,
+    cellsSupplementNotReplaceIsolation = true, placementDecisionsExplainable = true, unknownContaminationContainedAndEscalated = true,
+    selfServiceUsesSameControlsAsOperator = true, arch018AuthoritativeForTenantScope = true,
+    rebalancingChangesTenantIdentity = false, sharedInfrastructureMoveTreatedAsSimpleCostOptimization = false,
+    portabilityRequiresLowestCommonDenominator = false, policyAutomationResolvesAmbiguityAutonomously = false,
+    adaptiveControlsDiscriminateOnSensitiveAttributes = false, forecastsOverrideHealthOrContracts = false,
+    tenantDataFreeTrainingAsset = false, federatedComputationGuaranteesPrivacyAutomatically = false,
+    cryptographicIdentityProvesResourceAuthority = false, singleWorkflowFlagProvesErasure = false,
+    sustainabilityMovesDataOutsideApprovedLocations = false, experimentalCapabilityRedefinesMandatoryControls = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      evolutionPrinciples, stableArchitecturalInvariants, evolutionDrivers, evolutionHorizons, cellModelRequirements,
+      cellDesignOptions, placementEvaluationFactors, placementRebalancingSteps, progressiveIsolationRequirements,
+      sovereignDeploymentElements, providerPortabilityImprovements, futurePortabilityPackageContents,
+      continuousVerificationCapabilities, digitalTwinFields, policyAsCodeMaturityItems, adaptiveGovernanceTargets,
+      predictiveCapacityInputs, privacyPreservingControls, federatedLearningGovernanceAreas, aiAssistedOperationsAreas,
+      remediationDefinitionFields, selfServiceEvolutionItems, ecosystemGrowthAreas, collaborationEvolutionFields,
+      identityEvolutionPreservedElements, migrationFactoryMovementTypes, deletionVerificationDistinctions,
+      sustainabilityConsiderations, evolutionGovernanceSteps, evolutionMetrics, evolutionRisks, architecturalRules
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      tenantSemanticsStableAcrossEvolution, evolutionIncrementalReversibleEvidenceDriven, automationNarrowsRiskNotHidesDecisions,
+      isolationImprovementsPortable, dedicatedSovereignReuseCommonContracts, migrationIsProductCapability,
+      portabilityIncludesEvidence, aiOperationsBoundedByAccountability, privacyInsightsNoImplicitRights,
+      optimizationCannotWeakenCommitments, cellsSupplementNotReplaceIsolation, placementDecisionsExplainable,
+      unknownContaminationContainedAndEscalated, selfServiceUsesSameControlsAsOperator, arch018AuthoritativeForTenantScope,
+      rebalancingChangesTenantIdentity, sharedInfrastructureMoveTreatedAsSimpleCostOptimization,
+      portabilityRequiresLowestCommonDenominator, policyAutomationResolvesAmbiguityAutonomously,
+      adaptiveControlsDiscriminateOnSensitiveAttributes, forecastsOverrideHealthOrContracts, tenantDataFreeTrainingAsset,
+      federatedComputationGuaranteesPrivacyAutomatically, cryptographicIdentityProvesResourceAuthority,
+      singleWorkflowFlagProvesErasure, sustainabilityMovesDataOutsideApprovedLocations,
+      experimentalCapabilityRedefinesMandatoryControls
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}